refactor(webfont-generator): extract incremental SVG preparation
